// src/types/profile.ts
export interface ProfileFormData {
  profileName: string;
  displayName: string;
  bio: string;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  phone: string;
  email: string;
  address: string;
  city: string;
  state: string;
  zipCode: string;
  instagram: string;
  gender: string;
  profilePhotoUrl?: string;
}